Patch repr of HTTPRequest to not show body or sensitive HTTP headers
On stack traces, Tornado will log the repr of the HTTPRequest object
which had included POST body and if present, header for HTTP Basic
Auth. Both of these can potentially leak sensitive info into logs.

aside: It is ridiculous that this project inlines copies of two
different versions of Tornado yet also requires the Tornado package
to be installed.